==> 1-Data-Loading-Cleaning.ipynb <==
## <span style="color:#730101">Data Import and Cleaning</span>

In this notebook, we import the initial dataset from Yelp.com and perform the appropriate modifications to create the final data in order to continue the analysis. To be more specific, we clean the data, keep only the meaningfull columns and apply thorough text preprocessing.

##### <span style="color:#3A3A3A">Import Libraries</span>

import pandas as pd
import numpy as np
import json
import pickle
import glob
import os
import nltk
#nltk.download('words')
import re
import string 

from collections import Counter
from IPython.display import display
from typing import Tuple, List
from unidecode import unidecode
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
from nltk.corpus import wordnet
from ipywidgets import interact
from tqdm import tqdm, tqdm_notebook
tqdm_notebook().pandas()

import warnings
warnings.filterwarnings("ignore", 'This pattern has match groups')

pd.options.mode.chained_assignment = None

def set_pandas_display_options() -> None:
    """Set pandas display options."""
    display = pd.options.display
    display.max_columns = 1000
    display.max_rows = 1000
    display.max_colwidth = 199
    display.width = None
    # display.precision = 2  # set as needed

set_pandas_display_options()

##### <span style="color:#3A3A3A">Import Datasets</span>

'''
Import Business Dataset
-----------------------
The business dataset contains business data including location data, attributes, and categories.
'''

# Read the dataset in .json format
business = pd.read_json('yelp_academic_dataset_business.json', lines=True)
display(business.head(3))

# The business dataset has 209393 unique business_ids
print(business.shape)

'''
Import Review Dataset
---------------------
The Review Dataset contains full review text data including the user_id that wrote the review and the business_id the 
review is written for.
'''

# # Process the dataset in chunks of size 400000
# out_path = os.getcwd() #Path to save the pickle files to
# chunk_size = 400000 #size of chunks relies on your available memory

# reader = pd.read_json('yelp_academic_dataset_review.json', lines=True, chunksize=chunk_size)

# for i, chunk in enumerate(reader):
#     out_file = out_path + "/review_{}.pkl".format(i+1)
#     with open(out_file, "wb") as f:
#         pickle.dump(chunk,f,pickle.HIGHEST_PROTOCOL)      

pickle_path = os.getcwd() #Same Path as out_path i.e. where the pickle files are

data_p_files=[]
for name in glob.glob(pickle_path + "/review_*.pkl"):
       data_p_files.append(name)


df_reviews = pd.DataFrame([])
for i in range(len(data_p_files)):
    df_reviews = df_reviews.append(pd.read_pickle(data_p_files[i]),ignore_index=True)

display(df_reviews.head(3))

# The reviews dataset has 8021122 unique reviews_ids
df_reviews.shape

'''
Import User Dataset
-------------------
The User Dataset contains the user's friend mapping and all the metadata associated with the user.
'''

# # Process the dataset in chunks of size 400000
# out_path = os.getcwd() #Path to save the pickle files to
# chunk_size = 400000 #size of chunks relies on your available memory

# reader = pd.read_json('yelp_academic_dataset_user.json', lines=True, chunksize=chunk_size)

# for i, chunk in enumerate(reader):
#     out_file = out_path + "/user_{}.pkl".format(i+1)
#     with open(out_file, "wb") as f:
#         pickle.dump(chunk,f,pickle.HIGHEST_PROTOCOL)
        
pickle_path = os.getcwd() #Same Path as out_path i.e. where the pickle files are

data_p_files=[]
for name in glob.glob(pickle_path + "/user_*.pkl"):
       data_p_files.append(name)


df_user = pd.DataFrame([])
for i in range(len(data_p_files)):
    df_user = df_user.append(pd.read_pickle(data_p_files[i]),ignore_index=True)

display(df_user.head(3))

# The user dataset has 1968703 unique user_ids
df_user.shape

##### <span style="color:#3A3A3A">Data Cleaning</span>

# We keep the columns we want from the reviews and the user dataset in order to merge them based on the business_id.
# The aim is to create a dataframe in which we have the information about the business the user has made the review for.

# Keep columns from reviews dataset
reviews = df_reviews[['review_id','user_id','business_id','stars','text','date']] 
# Merge the reviews and business datasets
review_business = reviews.merge(business, left_on='business_id', right_on='business_id')

# Keep columns from user dataset
user = df_user[['user_id', 'name', 'review_count', 'average_stars']] 
# Merge the user, business and reviews datasets
restaurants = review_business.merge(user,left_on='user_id', right_on='user_id')
display(restaurants.head(3))

# We keep from the restaurants dataset the records that contain the word "restaurant" in its categories
restaurants['categories'] = restaurants['categories'].str.lower()
restaurants = restaurants[restaurants['categories'].str.contains('(restaurant).*')==True].reset_index()
# We also keep the restaurants that are open
restaurants = restaurants[restaurants['is_open']==1]

# We rename some columns in order to facilitate their use
restaurants = restaurants.rename(columns ={"stars_x":"review_stars"})
restaurants = restaurants.rename(columns ={"name_x":"business_name"})
restaurants = restaurants.rename(columns ={"stars_y":"rating"})
restaurants = restaurants.rename(columns ={"review_count_x":"num_of_business_reviews"})
restaurants = restaurants.rename(columns ={"name_y":"user_name"})
restaurants = restaurants.rename(columns ={"review_count_y":"num_of_reviews"})
restaurants.head(1)

# We also remove records in which the user has done more than one review in the same business and we keep only the 
# most recent one
restaurants=restaurants.sort_values('date', ascending=False).drop_duplicates(subset=['user_id', 'business_id'])
# The dataset, now, contains 4078659 records
print(restaurants.shape)

# We take a loon on the areas that are included in our dataset in order to decrease the records in our dataset.
group_restaur = restaurants[['state','text']]
group_restaur.groupby(['state']).count().sort_values(by=['text'], ascending=False)

# we keep only Nevada State as it contains enough records for our analysis
Nevada = restaurants[restaurants['state']=='NV']
Nevada = Nevada.drop('index',1)
Nevada = Nevada.reset_index(drop=True)
display(Nevada.head(3))

# The Nevada Dataset, now, contains 1362885 unique reviews.
Nevada.shape

##### <span style="color:#3A3A3A">Dataset Transformation</span>

# First, we are going to transform the cities in order to achieve a uniform appearance
Nevada['city'] = Nevada['city'].str.upper()
cities = Nevada.city
# Calculate the number of times the city appears
counts = cities.value_counts()
# Calculate number of unique businesses per city
unique_businesses =  Nevada.groupby('city')['business_id'].nunique()
# Calculate number of average reviews per city
avg_reviews = round((counts/unique_businesses),2)
table = pd.DataFrame({'counts':counts,'avg_reviews':avg_reviews,'unique_businesses':unique_businesses}).sort_values(by=['unique_businesses'], ascending=False)
display(table)

# Replace the cities with uniform names
Nevada['city'].replace(['N LAS VEGAS', 'N. LAS VEGAS', 'N.LAS VEGAS'], 'NORTH LAS VEGAS',inplace=True)
Nevada['city'].replace(['ENTERPRISE', 'SPRING VALLEY','4321 W FLAMINGO RD','SOUTH LAS VEGAS','SUMMERLIN','BLUE DIAMOND'], 'LAS VEGAS',inplace=True)
Nevada['city'].replace(['LAS  VEGAS'], 'LAS VEGAS',inplace=True)

# Delete these row indexes from dataframe as the contain little information
Nevada.drop(Nevada.loc[Nevada['city']=='SAN ANTONIO'].index, inplace=True)
Nevada.drop(Nevada.loc[Nevada['city']=='SUNRISE MANOR'].index, inplace=True)
Nevada.drop(Nevada.loc[Nevada['city']=='NELLIS AIR FORCE BASE'].index, inplace=True)
Nevada.drop(Nevada.loc[Nevada['city']=='NELLIS AFB'].index, inplace=True)

# We have another look on the cities now
cities = Nevada.city
# Calculate the number of times the city appears
counts = cities.value_counts()
# Calculate number of unique businesses per city
unique_businesses =  Nevada.groupby('city')['business_id'].nunique()
# Calculate number of average reviews per city
avg_reviews = round((counts/unique_businesses),2)
table = pd.DataFrame({'counts':counts,'avg_reviews':avg_reviews,'unique_businesses':unique_businesses}).sort_values(by=['unique_businesses'], ascending=False)
table

# The Nevada Dataset, now, contains 1362729 unique reviews.
Nevada.shape

## <span style="color:#730101">Text Processing</span>

##### <span style="color:#3A3A3A">Lower Text</span>

# Create new column in which the text process will be applied. 
Nevada["cleanText"] = Nevada["text"]

# Convert to lowercase
Nevada.cleanText = Nevada.cleanText.str.lower()

##### <span style="color:#3A3A3A">Remove non-english words and symbols</span>

# We discovered that our dataset also contained non-english characters, like chinese characters

words = set(nltk.corpus.words.words())

def eng_words(text):
    new=" ".join(w for w in nltk.wordpunct_tokenize(text) 
         if w.lower() in words or not w.isalpha())
    eng = re.sub("[^a-zA-Z]+", " ", new)
    return eng

Nevada['cleanText'] = Nevada['cleanText'].progress_apply(eng_words)

##### <span style="color:#3A3A3A">Contractions</span>

# Based on https://stackoverflow.com/questions/43018030/replace-apostrophe-short-words-in-python
# we are going to replace the apostrophe in short words in our text.

import re

def decontracted(phrase):
    # specific
    phrase = re.sub(r"won\'t", "will not", phrase)
    phrase = re.sub(r"can\'t", "can not", phrase)

    # general
    phrase = re.sub(r"n\'t", " not", phrase)
    phrase = re.sub(r"\'re", " are", phrase)
    phrase = re.sub(r"\'s", " is", phrase)
    phrase = re.sub(r"\'d", " would", phrase)
    phrase = re.sub(r"\'ll", " will", phrase)
    phrase = re.sub(r"\'t", " not", phrase)
    phrase = re.sub(r"\'ve", " have", phrase)
    phrase = re.sub(r"\'m", " am", phrase)
    
    return phrase

Nevada['cleanText'] = Nevada['cleanText'].progress_apply(decontracted)

##### <span style="color:#3A3A3A">Remove Punctuation</span>

def remove_punctuations(text):
    for punctuation in string.punctuation:
        text = text.replace(punctuation, '')
    return text

Nevada['cleanText'] = Nevada['cleanText'].progress_apply(remove_punctuations)

##### <span style="color:#3A3A3A">Convert Accented Characters</span>

# Based on the insights of https://towardsdatascience.com/nlp-text-preprocessing-a-practical-guide-and-template-d80874676e79
# we are converting assected characters like café

def remove_accented_chars(text):
    """remove accented characters from text, e.g. café"""
    text = unidecode(text)
    return text

Nevada['cleanText'] = Nevada['cleanText'].progress_apply(remove_accented_chars)

##### <span style="color:#3A3A3A">Remove Whitespace</span>

def remove_whitespace(text):
  text = text.strip()
  return text

Nevada['cleanText'] = Nevada['cleanText'].progress_apply(remove_whitespace)   

##### <span style="color:#3A3A3A">Tokenization</span>

# The next step is to tokenize the reviews text. For this reason, we are using word_tokenize from NLTK.
# Basically, we will split the text into word tokens while we will also remove any words that are only one character long. 
# Finally, we remove numbers, but not words that contain numbers.

# Create new column in which the text process will be applied. 
Nevada["tokenText"]=Nevada["cleanText"]

def clean_token(text):
  return word_tokenize(text)

Nevada["tokenText"] = Nevada["tokenText"].progress_apply(clean_token)

#Remove words that are only one character.
Nevada["tokenText"] = [[token for token in doc if len(token) > 1] for doc in Nevada["tokenText"]]

# Remove numbers, but not words that contain numbers.
Nevada["tokenText"] = [[token for token in doc if not token.isnumeric()] for doc in Nevada["tokenText"]]

display(Nevada.tokenText)

##### <span style="color:#3A3A3A">Stopwords</span>

# Another step in the text cleaning process is the removal of stopwords. Basically, stopwords are a set of commonly 
# used words like "a", "the", "is" and etc. The purpose behind the removal of stopwords is that by removing low information
# words from a text, we can genuinely focus on important words instead.

stopwords = (['ll',  'namely', 'fifty', 'sure', 'bottom', 'rd', 'indicate', 'usefully', 'anyway', 'or', 'haven', 'thorough', 'different', 'she', 'looking', 'whim',
 've', 'nay', 'fify', 'system', 'brief', 're', 'indicated', 'usefulness', 'anyways', 'ord', 'havent', 'thoroughly', 'do', 'shed', 'looks', 'whither',
 'a', 'nd', 'fill', 't', 'briefly', 'readily', 'indicates', 'uses', 'anywhere', 'other', 'having', 'those', 'does', 'shell', 'ltd', 'who',
 'able', 'near', 'find', 'take', 'but', 'really', 'information', 'using', 'apart', 'others', 'he', 'thou', 'doesn', 'shes', 'm', 'whod',
 'about', 'nearly', 'fire', 'taken', 'by', 'reasonably', 'inner', 'usually', 'apparently', 'otherwise', 'hed', 'though', 'doesnt', 'should', 'ma', 'whoever',
 'above', 'necessarily', 'first', 'taking', 'c', 'recent', 'insofar', 'v', 'appear', 'ought', 'hell', 'thoughh', 'doing', 'shouldn', 'made', 'whole',
 'abst', 'necessary', 'five', 'tell', "c'mon", 'recently', 'instead', 'value', 'appreciate', 'our', 'hello', 'thousand', 'don', 'shouldnt', 'mainly', 'wholl',
 'accordance', 'need', 'fix', 'ten', "c's", 'ref', 'interest', 'various', 'appropriate', 'ours', 'help', 'three', "don't", 'shouldve', 'make', 'whom',
 'according', 'needn', 'followed', 'tends', 'ca', 'refs', 'into', 've', 'approximately', 'ourselves', 'hence', 'throug', 'done', 'show', 'makes', 'whomever',
 'accordingly', "needn't", 'following', 'th', 'call', 'regarding', 'invention', 'very', 'are', 'out', 'her', 'through', 'down', 'showed', 'many', 'whos',
 'across', 'needs', 'follows', 'than', 'came', 'regardless', 'inward', 'via', 'aren', 'outside', 'here', 'throughout', 'downwards', 'shown', 'may', 'whose',
 'act', 'neither', 'for', 'thank', 'can', 'regards', 'is', 'viz', 'arent', 'over', 'hereafter', 'thru', 'due', 'showns', 'maybe', 'why',
 'actually', 'never', 'former', 'thanks', 'cannot', 'related', 'isn', 'vol', 'arise', 'overall', 'hereby', 'thus', 'during', 'shows', 'me', 'whys',
 'added', 'nevertheless', 'formerly', 'thanx', 'cant', 'relatively', 'isnt', 'vols', 'around', 'owing', 'herein', 'til', 'e', 'side', 'mean', 'widely',
 'adj', 'new', 'forth', 'that', 'cause', 'research', 'it', 'vs', 'as', 'own', 'heres', 'tip', 'each', 'significant', 'means', 'will',
 'affected', 'next', 'forty', "that'll", 'causes', 'respectively', 'itd', 'w', 'aside', 'p', 'hereupon', 'to', 'ed', 'significantly', 'meantime', 'willing',
 'affecting', 'nine', 'found', 'thats', 'certain', 'resulted', 'itll', 'want', 'ask', 'page', 'hers', 'today', 'edu', 'similar', 'meanwhile', 'wish',
 'affects', 'ninety', 'four', 'thatve', 'certainly', 'resulting', 'its', 'wants', 'asking', 'pages', 'herself', 'together', 'effect', 'similarly', 'merely', 'with',
 'after', 'nobody', 'from', 'the', 'changes', 'results', 'itself', 'was', 'associated', 'part', 'hes', 'tomorrow', 'eg', 'since', 'mg', 'within',
 'afterwards', 'non', 'front', 'their', 'clearly', 'right', 'ive', 'wasn', 'at', 'particular', 'hi', 'too', 'eight', 'sincere', 'might', 'without',
 'again', 'none', 'full', 'theirs', 'co', 'run', 'j', 'wasnt', 'auth', 'particularly', 'hid', 'took', 'eighty', 'six', 'mightn', 'won',
 'against', 'nonetheless', 'further', 'them', 'com', 's', 'just', 'way', 'available', 'past', 'him', 'top', 'either', 'sixty', 'mightnt', 'wonder',
 'ah', 'noone', 'furthermore', 'themselves', 'come', 'said', 'k', 'we', 'away', 'per', 'himself', 'toward', 'eleven', 'slightly', 'mill', 'wont',
 'ain', 'nor', 'g', 'then', 'comes', 'same', 'keep', 'wed', 'awfully', 'perhaps', 'his', 'towards', 'else', 'so', 'million', 'words',
 'aint', 'normally', 'gave', 'thence', 'con', 'saw', 'keeps', 'welcome', 'b', 'placed', 'hither', 'tried', 'elsewhere', 'some', 'mine', 'world',
 'all', 'nos', 'get', 'there', 'concerning', 'say', 'kept', 'well', 'back', 'please', 'home', 'tries', 'empty', 'somebody', 'miss', 'would',
 'allow', 'noted', 'gets', 'thereafter', 'consequently', 'saying', 'kg', 'went', 'be', 'plus', 'hopefully', 'truly', 'end', 'somehow', 'ml', 'wouldn',
 'allows', 'nothing', 'getting', 'thereby', 'consider', 'says', 'km', 'were', 'became', 'poorly', 'how', 'try', 'ending', 'someone', 'more', 'wouldnt',
 'almost', 'novel', 'give', 'thered', 'considering', 'sec', 'know', 'weren', 'because', 'possible', 'howbeit', 'trying', 'enough', 'somethan', 'moreover', 'www',
 'alone', 'now', 'given', 'therefore', 'contain', 'second', 'known', 'werent', 'become', 'possibly', 'however', 'ts', 'entirely', 'something', 'morning', 'x',
 'along', 'nowhere', 'gives', 'therein', 'containing', 'secondly', 'knows', 'weve', 'becomes', 'potentially', 'hows', 'twelve', 'especially', 'sometime', 'most', 'y',
 'already', 'o', 'giving', 'therell', 'contains', 'section', 'l', 'what', 'becoming', 'pp', 'http', 'twenty', 'et', 'sometimes', 'mostly', 'yes',
 'also', 'obtain', 'go', 'thereof', 'corresponding', 'see', 'largely', 'whatever', 'been', 'predominantly', 'https', 'twice', 'etc', 'somewhat', 'move', 'yet',
 'although', 'obtained', 'goes', 'therere', 'could', 'seeing', 'last', 'whatll', 'before', 'present', 'hundred', 'two', 'even', 'somewhere', 'mr', 'you',
 'always', 'obviously', 'going', 'theres', 'couldn', 'seem', 'lately', 'whats', 'beforehand', 'presumably', 'i', 'u', 'ever', 'soon', 'mrs', 'youd',
 'am', 'of', 'gone', 'thereto', 'couldnt', 'seemed', 'later', 'when', 'begin', 'previously', 'id', 'un', 'every', 'sorry', 'much', 'youll',
 'among', 'off', 'got', 'thereupon', 'course', 'seeming', 'latter', 'whence', 'beginning', 'primarily', 'ie', 'under', 'everybody', 'specifically', 'mug', 'your',
 'amongst', 'often', 'gotten', 'thereve', 'cry', 'seems', 'latterly', 'whenever', 'beginnings', 'probably', 'if', 'unfortunately', 'everyone', 'specified', 'must', 'youre',
 'amoungst', 'oh', 'greetings', 'these', 'currently', 'seen', 'least', 'whens', 'begins', 'promptly', 'ignored', 'unless', 'everything', 'specify', 'mustn', 'yours',
 'amount', 'ok', 'h', 'they', 'd', 'self', 'less', 'where', 'behind', 'proud', 'ill', 'unlike', 'everywhere', 'specifying', 'mustnt', 'yourself',
 'an', 'okay', 'had', 'theyd', 'date', 'selves', 'lest', 'whereafter', 'being', 'provides', 'im', 'unlikely', 'ex', 'still', 'my', 'yourselves',
 'and', 'old', 'hadn', 'theyll', 'de', 'sensible', 'let', 'whereas', 'believe', 'put', 'immediate', 'until', 'exactly', 'stop', 'myself', 'yourselvesme',
 'announce', 'omitted', "hadn't", 'theyre', 'definitely', 'sent', 'lets', 'whereby', 'below', 'q', 'immediately', 'unto', 'example', 'strongly', 'n', 'youve',
 'another', 'on', 'happens', 'theyve', 'describe', 'serious', 'like', 'wherein', 'beside', 'que', 'importance', 'up', 'except', 'sub', 'na', 'z',
 'any', 'once', 'hardly', 'thick', 'described', 'seriously', 'liked', 'wheres', 'besides', 'quickly', 'important', 'upon', 'f', 'substantially', 'name', 'zero',
 'anybody', 'one', 'has', 'thickv', 'despite', 'seven', 'likely', 'whereupon', 'between', 'quite', 'in', 'ups', 'far', 'successfully',
 'anyhow', 'ones', 'hasn', 'thin', 'detail', 'several', 'line', 'wherever', 'beyond', 'qv', 'inasmuch', 'us', 'few', 'such',
 'anymore', 'only', "hasn't", 'think', 'did', 'shall', 'little', 'whether', 'bill', 'r', 'inc', 'use', 'ff', 'sufficiently',
 'anyone', 'onto', 'hasnt', 'third', 'didn', 'shan', 'll', 'which', 'biol', 'ran', 'indeed', 'used', 'fifteen', 'suggest',
 'anything', 'opa', 'have', 'this', "didn't", 'shant', 'look', 'while', 'both', 'rather', 'index', 'useful', 'fifth', 'sup'
])

Nevada["tokenText"] = [[token for token in doc if (token not in stopwords)] for doc in Nevada["tokenText"]]

display(Nevada.head(3))

##### <span style="color:#3A3A3A">Lemmatization</span>

# Lemmatization is another normalization technique and basically is the conversion of each word to its base form, 
# the lemma that we encounter in dictionaries. We are lemmatising our text based on 
# https://simonhessner.de/lemmatize-whole-sentences-with-python-and-nltks-wordnetlemmatizer/

# Create new column in which the Lemmatization process will be applied. 
Nevada["LemaText"]=Nevada["tokenText"]

lemmatizer = WordNetLemmatizer()
def nltk2wn_tag(nltk_tag):
  if nltk_tag.startswith('J'):
    return wordnet.ADJ
  elif nltk_tag.startswith('V'):
    return wordnet.VERB
  elif nltk_tag.startswith('N'):
    return wordnet.NOUN
  elif nltk_tag.startswith('R'):
    return wordnet.ADV
  else:                    
    return None

def lemmatize_sentence(sentence):
    nltk_tagged = nltk.pos_tag((sentence))   
    wn_tagged = map(lambda x: (x[0], nltk2wn_tag(x[1])), nltk_tagged)
    res_words = []
    for word, tag in wn_tagged:
        if tag is None:                        
            res_words.append(word)
        else:
            res_words.append(lemmatizer.lemmatize(word, tag))
    return " ".join(res_words)

Nevada["LemaText"] = Nevada["LemaText"].progress_apply(lemmatize_sentence)

display(Nevada.LemaText)

Nevada = Nevada[Nevada.LemaText.str.len()!=0]

Nevada['split_text'] = Nevada['LemaText'].str.split()
# Remove records that contain no word 
Nevada = Nevada[Nevada.split_text.str.len()!=0]
# Remove records that contain only one word
Nevada = Nevada[Nevada.split_text.str.len()!=1]

# Get the average length of a review
round(Nevada.split_text.str.len().mean(),2)

# Get the max length of a review
round(Nevada.split_text.str.len().max(),2)

Nevada = Nevada.drop(columns=['split_text'])

Nevada.shape

##### <span style="color:#3A3A3A">Treat Categories</span>

# Create a new column where the text processing for the categories will be applied
Nevada['cleanCat'] = Nevada['categories']

# Remove Punctuations
Nevada['cleanCat'] = Nevada['cleanCat'].progress_apply(remove_punctuations)

# Remove Extra Space
Nevada['cleanCat'] = [text.replace("  "," ") for text in Nevada['cleanCat']]

display(Nevada['cleanCat'])

# Get a view on the categories and their frequency on our dataset
most_common = Counter(" ".join(Nevada['cleanCat']).split(" ")).most_common()
most_common

#only get 23 most common cuisine categories
Nevada = Nevada[(Nevada['cleanCat'].str.contains('american|mexican|japanese|italian|asian|chinese|korean|french|mediterranean|hawaiian|vietnamese|greek|ethnic|spanish|brazilian|taiwanese|african|british|caribbean|pakistani|thai|indian|irish',case=False))].copy()


#We decide to keep only get 23 most common cuisine categories for our analysis

cuisine = ['american', 'mexican', 'japanese', 'italian', 'asian', 'chinese', 'korean', 'french', 'mediterranean', 'hawaiian', 'vietnamese', 'greek', 'ethnic', 'spanish', 'brazilian', 'taiwanese', 'african', 'british', 'caribbean', 'pakistani', 'thai', 'indian', 'irish']   
Nevada['cuisines'] = Nevada.cleanCat.apply(lambda x: ' '.join([word for word in x.split() if word in (cuisine)]))

# Remove duplicates cuisines categories

def uniquify(string):
    output = []
    seen = set()
    for word in string.split():
        if word not in seen:
            output.append(word)
            seen.add(word)
    return ' '.join(output)

Nevada['cuisines'] = Nevada['cuisines'].apply(uniquify)

# Remove Whitespace
def remove_whitespace(text):
  text = text.strip()
  return text

Nevada['cuisines'] = Nevada['cuisines'].apply(remove_whitespace)   

# Get a view on the cuisines and their frequency on our dataset.
# The american is expected to be the most frequent cuisine of the businesses.
most_common = Counter(" ".join(Nevada['cuisines']).split(" ")).most_common()
most_common

# Merge the clean review text and the categories into one column in order to enhance the information of our dataset.
Nevada['input_text'] = Nevada[['LemaText', 'cuisines']].agg(' '.join, axis=1)

# Drop columns that we will not use in our analysis
Nevada = Nevada.drop(columns=['num_of_business_reviews', 'is_open','hours','num_of_reviews','cleanText','tokenText', 'cleanCat']).reset_index(drop=True)

display(Nevada.head(3))

# The final dataset contains 1013794 unique reviews.
Nevada.shape

file_name="Nevada.pkl"
Nevada.to_pickle(file_name)
