==> 5-Recommendation-System.ipynb <==
## <span style="color:#730101">Classification of Review Stars</span>

In this entity, we create the recommendation system.

#### Import Modules

import pandas as pd
import numpy as np
import os
import pickle
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras.models import Model

pd.options.mode.chained_assignment = None

self_trained_model_cnn = keras.models.load_model(os.path.join(os.getcwd(), 'Self_Trained_Model_CNN'))

# We select the best model and we rename it to model in order to continue the process
model = self_trained_model_cnn

### Extract embeddings

model.layers

You can think of embedding layer as a lookup table where nth row of the table correspond to word vector of the nth word (but embedding layers is trainable layer not just a static lookup table)

# Import Dataset
with open('Nevada.pkl', 'rb') as nevada:
    Nevada = pickle.load(nevada)

Nevada = Nevada.drop(columns=['attributes','categories','LemaText', ]).reset_index(drop=True)
Nevada.head(2)

Nevada.shape

### Recommendation System

#### Step 1: Get the user's id and keep his reviews

# The user has to input his/her id/username
input_userid = input("Give me your id: ") 
# Test this one user_id with lots of reviews: iW6YSCu3YVI-SNPNi0I-xg  
# another user_id with only one review: f38DGhAQOxQu07P-4JMpTw

# Keep the data for the user of interest
user_data = Nevada[Nevada.user_id==input_userid]
user_data

user_data.shape

# Get User's Name
name = user_data.user_name.unique()
print(f'Hello {name[0]}.\n')

# Sort the dataset by review_star and date of review and get the first review.
user_data = user_data.sort_values(['review_stars', 'date'], ascending=[False, False]).reset_index(drop=True)
user_review = user_data.input_text[0]
user_review

#### Step 2 Run the Recommendation Function

# Extract embeddings
text_weights = model.layers[1].get_weights()[0]
print("The shape of embedded weights: ", text_weights.shape)
print("The length of embedded weights: ", len(text_weights))

# We need to normalize the embeddings so that the dot product between two embeddings becomes the cosine similarity.
text_weights = text_weights / np.linalg.norm(text_weights, axis = 1).reshape((-1, 1))
round(np.sum(np.square(text_weights[0])))

texts = Nevada['input_text']

review_index = {link: idx for idx, link in enumerate(texts)}
index_review = {idx: link for link, idx in review_index.items()}

review_index[user_review]

* We have based the below function on this [article](https://www.kaggle.com/willkoehrsen/neural-network-embedding-recommendation-system)

# The function below takes in either a user review, a set of embeddings, and returns the n most similar items to the 
# review. It does this by computing the dot product between the review and embeddings. Because we normalized the embeddings, 
# the dot product represents the cosine similarity between two vectors. Once we have the dot products, we can sort the 
# results to find the closest entities in the embedding space. With cosine similarity, higher numbers indicate entities 
# that are closer together, with -1 the furthest apart and +1 closest together.

def find_similar(name, weights, index_name = 'review', n = 20, return_dist = False):
    """Find n most similar items (or least) to name based on embeddings. Option to also plot the results"""
    
    # Select index and reverse index
    if index_name == 'review':
        index = review_index
        rindex = index_review
    
    # Check to make sure `name` is in index
    try:
        # Calculate dot product between review and all others
        dists = np.dot(weights, weights[index[name]])
    except KeyError:
        print(f'{name} Not Found.')
        return
    
    # Sort distance indexes from smallest to largest
    sorted_dists = np.argsort(dists)
    
    # Take the last n sorted distances
    closest = sorted_dists[-n:]

    # Need distances later on
    if return_dist:
        return dists, closest


    print(f'{index_name.capitalize()}s closest to {name}.\n')
        
    # Need distances later on
    if return_dist:
        return dists, closest
    
    
    # Print formatting
    max_width = max([len(rindex[c]) for c in closest])
    
    review_text=[]
    similarity=[]
    # Print the most similar and distances
    for c in reversed(closest):
        #print(f'{index_name.capitalize()}: {rindex[c]:{max_width + 2}} Similarity: {dists[c]:.{2}}')
        review_text.append(rindex[c])
        similarity.append(round(dists[c],3))
    data_similar = pd.DataFrame(np.column_stack([review_text, similarity]),
                                columns=['Review', 'Similarity'])
    
    return data_similar

# Print the top 20 most similar Reviews
pd.set_option('max_colwidth', 300)
data_similar = find_similar(user_review, text_weights)
data_similar

#### Step 3 Recommend Restaurant

restaurant_name=[]
restaurant_cuisine=[]
restaurant_stars=[]
user=[]

for i in range(len(data_similar)):
    if i==0:
        continue
    else:
        # Create the list of the reaturant's names
        res_name = Nevada[Nevada.input_text==data_similar.Review[i]].reset_index(drop=True).business_name
        res_name = res_name[0]
        restaurant_name.append(res_name)
        # Create the list of the reaturant's cuisine
        res_cuisine = Nevada[Nevada.input_text==data_similar.Review[i]].reset_index(drop=True).cuisines
        res_cuisine = res_cuisine[0]
        restaurant_cuisine.append(res_cuisine)
        # Create the list of the reaturant's average stars rating
        res_stars = Nevada[Nevada.input_text==data_similar.Review[i]].reset_index(drop=True).average_stars
        res_stars = round(res_stars[0],3)
        restaurant_stars.append(res_stars)
        # Create the list of the reaturant's user id for the specific review
        res_user = Nevada[Nevada.input_text==data_similar.Review[i]].reset_index(drop=True).user_id
        res_user = res_user[0]
        user.append(res_user)

# Create the dataframe that holds the recommendations
recommendations = pd.DataFrame(np.column_stack([restaurant_name, restaurant_cuisine, restaurant_stars,user]),
                               columns=['Name', 'Cuisine', 'Stars', 'User'])

recommendations = recommendations[recommendations.User!=input_userid]
recommendations

input_cuisine = input("What cuisine would you like to try today? ").lower() #put as input: american or mexican

# Get the name of the chosen restaurant
Name = recommendations[recommendations.Cuisine.str.contains(input_cuisine)].reset_index(drop=True).Name[0]
# Get the location of the chosen restaurant
Location = Nevada[Nevada.business_name==Name].reset_index(drop=True).address[0]
print(f'You should try {Name} at {Location}.\n')
print(f'Have a nice meal!\n')

## Restaurant Recommendation

# Extract embeddings
text_weights = model.layers[1].get_weights()[0]
print("The shape of embedded weights: ", text_weights.shape)
print("The length of embedded weights: ", len(text_weights))

# We need to normalize the embeddings so that the dot product between two embeddings becomes the cosine similarity.
text_weights = text_weights / np.linalg.norm(text_weights, axis = 1).reshape((-1, 1))
round(np.sum(np.square(text_weights[0])),1)

texts = Nevada['input_text']

review_index = {link: idx for idx, link in enumerate(texts)}
index_review = {idx: link for link, idx in review_index.items()}

def RestaurantRecommendation(weights, review_index,index_review,index_name = 'review',n=20 ):
    
    input_userid = input("Give me your id: ")
    
    user_data = Nevada[Nevada.user_id==input_userid]
    
    # Get User's Name
    name = user_data.user_name.unique()
    print(f'\nHello {name[0]}.\n')
    
    # Sort the dataset by review_star and date of review and get the first review.
    user_data = user_data.sort_values(['review_stars', 'date'], ascending=[False, False]).reset_index(drop=True)
    user_review = user_data.input_text[0]

    # Select index and reverse index
    if index_name == 'review':
        index = review_index
        rindex = index_review
    
    # Check to make sure `user_review` is in index
    try:
        # Calculate dot product between review and all others
        dists = np.dot(weights, weights[index[user_review]])
    except KeyError:
        print(f'{user_review} Not Found.')
        return
    
    # Sort distance indexes from smallest to largest
    sorted_dists = np.argsort(dists)
    
    # Take the last n sorted distances
    closest = sorted_dists[-n:]


    #print(f'{index_name.capitalize()}s closest to {user_review}.\n')    
    
    # Print formatting
    max_width = max([len(rindex[c]) for c in closest])
    
    review_text=[]
    similarity=[]
    # Print the most similar and distances
    for c in reversed(closest):
        #print(f'{index_name.capitalize()}: {rindex[c]:{max_width + 2}} Similarity: {dists[c]:.{2}}')
        review_text.append(rindex[c])
        similarity.append(dists[c])
    data_similar = pd.DataFrame(np.column_stack([review_text, similarity]),
                                columns=['Review', 'Similarity'])
    
    restaurant_name=[]
    restaurant_cuisine=[]
    restaurant_stars=[]
    user=[]
    for i in range(len(data_similar)):
        if i==0:
            continue
        else:
            res_name = Nevada[Nevada.input_text==data_similar.Review[i]].reset_index(drop=True).business_name
            res_name = res_name[0]
            restaurant_name.append(res_name)
            res_cuisine = Nevada[Nevada.input_text==data_similar.Review[i]].reset_index(drop=True).cuisines
            res_cuisine = res_cuisine[0]
            restaurant_cuisine.append(res_cuisine)
            res_stars = Nevada[Nevada.input_text==data_similar.Review[i]].reset_index(drop=True).average_stars
            res_stars = round(res_stars[0],3)
            restaurant_stars.append(res_stars)
            res_user = Nevada[Nevada.input_text==data_similar.Review[i]].reset_index(drop=True).user_id
            res_user = res_user[0]
            user.append(res_user)

    recommendations = pd.DataFrame(np.column_stack([restaurant_name, restaurant_cuisine, restaurant_stars,user]),
                               columns=['Name', 'Cuisine', 'Stars', 'User'])
    recommendations = recommendations[recommendations.User!=input_userid]
    input_cuisine = input("What cuisine would you like to try today? ").lower()
    
    Name = recommendations[recommendations.Cuisine.str.contains(input_cuisine)].reset_index(drop=True).Name[0]
    Location = Nevada[Nevada.business_name==Name].reset_index(drop=True).address[0]
    print(f'You should try {Name} at {Location}.\n')
    print(f'Have a nice meal!\n')

RestaurantRecommendation(text_weights, review_index, index_review)
